--- backend/response_generator.py
@@ -1,6 +1,0 @@
-# backend/response_generator.py
-
-def generate_response(user_message):
-    print(f"User: {user_message}")
-    bot_response = input("Your response: ")
-    return bot_response
